+++ cmd/hidden_lake/traffic/internal/handler/pointer.go
@@ -11,7 +11,7 @@ import (
 	"github.com/number571/go-peer/pkg/logger"
 )
 
-func HandlePointerAPI(pWrapperDB database.IWrapperDB, pLogger logger.ILogger) http.HandlerFunc {
+func HandlePointerAPI(pDBWrapper database.IDBWrapper, pLogger logger.ILogger) http.HandlerFunc {
 	return func(pW http.ResponseWriter, pR *http.Request) {
 		logBuilder := http_logger.NewLogBuilder(hlt_settings.CServiceName, pR)
 
@@ -21,7 +21,7 @@ func HandlePointerAPI(pWrapperDB database.IWrapperDB, pLogger logger.ILogger) ht
 			return
 		}
 
-		database := pWrapperDB.Get()
+		database := pDBWrapper.Get()
 		if database == nil {
 			pLogger.PushErro(logBuilder.WithMessage("get_database"))
 			api.Response(pW, http.StatusInternalServerError, "failed: get database")