+++ pkg/network/anonymity/examples/ping-ping/main.go
@@ -140,7 +140,7 @@ func handler(serviceName string) anonymity.IHandlerF {
 
 func closeNode(node anonymity.INode) error {
 	return utils.MergeErrors(
-		node.GetWrapperDB().Close(),
+		node.GetDBWrapper().Close(),
 		node.GetNetworkNode().Close(),
 	)
 }
@@ -183,7 +183,7 @@ func newNode(serviceAddress, name, dbPath string) anonymity.INode {
 				)
 			},
 		),
-		anonymity.NewWrapperDB().Set(db),
+		anonymity.NewDBWrapper().Set(db),
 		network.NewNode(
 			nodeSettings(serviceAddress),
 			queue_set.NewQueueSet(